+++ partiql-lang/src/main/kotlin/org/partiql/lang/prettyprint/QueryPrettyPrinter.kt
@@ -333,7 +333,7 @@ class QueryPrettyPrinter {
     }
 
     private fun writeSessionAttribute(node: PartiqlAst.Expr.SessionAttribute, sb: StringBuilder) {
-        sb.append(node.value.text)
+        sb.append(node.value.text.uppercase())
     }
 
     /**