+++ CHANGELOG.md
@@ -32,8 +32,11 @@ Thank you to all who have contributed!
 - Adds `org.partiql.value` (experimental) package for reading/writing PartiQL
   values
 - Adds serialization and deserialization between IonValue and `org.partiql.value`.
+- Adds `org.partiql.ast` package and usage documentation
+- Adds `org.partiql.parser` package and usage documentation
 - Adds PartiQL's Timestamp Data Model. 
 - Adds support for Timestamp constructor call in Parser. 
+
 
 ### Changed